requirements.txt
Django
whitenoise
azure-iot-device
azure.iot.hub